Разделяет проект на модули и обозначает файловую структуру
