+++ js/request.js
@@ -1,0 +1,1 @@
+// Отправка данных на сервер (AJAX)
